release(0.3): substantially improve API ergonomics and remove build-time bindgen

+++ spirv-to-dxil/src/error.rs
@@ -1,4 +1,13 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+/// Error type for spirv-to-dxil.
 pub enum SpirvToDxilError {
+    /// An error occurred when compiling SPIR-V to DXIL.
+    #[error("An error occurred when compiled to DXIL: {0}.")]
     CompilerError(String),
-    RegisterSpaceOverflow(usize)
+    /// The requested register space for a push or runtime constant block is beyond
+    /// the limit of 31.
+    #[error("Register space {0} is beyond the limit of 31.")]
+    RegisterSpaceOverflow(u32)
 }
